Database.Beam.Sqlite.Connection

Documentation

data Sqlite

The SQLite backend. Used to parameterize MonadBeam and FromBackendRow to provide support for SQLite databases. See the documentation for MonadBeam and the user guide for more information on how to use this backend.

Constructors

Sqlite

newtype SqliteM a

MonadBeam instance inside which SQLite queries are run. See the user guide for more information

Constructors

SqliteM

Fields

runSqliteM :: ReaderT (String -> IO (), Connection) IO a

Run an IO action with access to a SQLite connection and a debug logging function, called or each query submitted on the connection.

sqliteUriSyntax :: c Sqlite Connection SqliteM -> BeamURIOpeners c

URI syntax for use with withDbConnection. See documentation for BeamURIOpeners for more information.

runBeamSqlite :: Connection -> SqliteM a -> IO a

runBeamSqliteDebug :: (String -> IO ()) -> Connection -> SqliteM a -> IO a

Emulated INSERT RETURNING support

insertReturning :: Beamable table => DatabaseEntity Sqlite db (TableEntity table) -> SqlInsertValues Sqlite (table (QExpr Sqlite s)) -> SqlInsert Sqlite table

Build a SqliteInsertReturning representing inserting the given values into the given table. Use runInsertReturningList

runInsertReturningList :: (Beamable table, FromBackendRow Sqlite (table Identity)) => SqlInsert Sqlite table -> SqliteM [table Identity]

Runs a SqliteInsertReturning statement and returns a result for each inserted row.