Database.Beam.Sqlite.Syntax

Description

SQLite implementations of the Beam SQL syntax classes

The SQLite command syntax is implemented by SQLiteCommandSyntax.

SQLite syntaxes

data SqliteSyntax

The syntax for SQLite is stored as a Builder along with a list of data that hasn't been serialized yet.

The first argument is a function that receives a builder for SQLData and returns the concrete syntax to embed into the query. For queries sent to the backend, this is simply a function that returns "?". Thus, the syntax sent to the backend includes proper placeholders. The list of data is sent to the SQLite library for proper escaping.

When the syntax is being serialized for display (for use in beam migrate for example), the data builder attempts to properly format and escape the data. This returns syntax suitable for inclusion in scripts. In this case, the value list is ignored.

Constructors

SqliteSyntax ((SQLData -> Builder) -> Builder) (DList SQLData)

data SqliteCommandSyntax

A SQLite command. INSERT is special cased to handle AUTO INCREMENT columns. The fromSqliteCommand function will take an SqliteCommandSyntax and convert it into the correct SqliteSyntax.

Constructors

SqliteCommandSyntax SqliteSyntax

SqliteCommandInsert SqliteInsertSyntax

newtype SqliteSelectSyntax

SQLite SELECT syntax

Constructors

SqliteSelectSyntax

Fields

fromSqliteSelect :: SqliteSyntax

data SqliteInsertSyntax

SQLite INSERT syntax. This doesn't directly wrap SqliteSyntax because we need to do some processing on INSERT statements to deal with AUTO INCREMENT columns. Use formatSqliteInsert to turn SqliteInsertSyntax into SqliteSyntax.

Constructors

SqliteInsertSyntax

Fields

sqliteInsertTable :: !SqliteTableNameSyntax

sqliteInsertFields :: [Text]

sqliteInsertValues :: !SqliteInsertValuesSyntax

sqliteInsertOnConflict :: !(Maybe SqliteOnConflictSyntax)

newtype SqliteUpdateSyntax

SQLite UPDATE syntax

Constructors

SqliteUpdateSyntax

Fields

fromSqliteUpdate :: SqliteSyntax

newtype SqliteDeleteSyntax

SQLite DELETE syntax

Constructors

SqliteDeleteSyntax

Fields

fromSqliteDelete :: SqliteSyntax

newtype SqliteOnConflictSyntax

SQLite ON CONFLICT syntax

Constructors

SqliteOnConflictSyntax

Fields

fromSqliteOnConflict :: SqliteSyntax

data SqliteInsertValuesSyntax

SQLite VALUES clause in INSERT. Expressions need to be handled explicitly in order to deal with DEFAULT values and AUTO INCREMENT columns.

Constructors

SqliteInsertExpressions [[SqliteExpressionSyntax]]

SqliteInsertFromSql SqliteSelectSyntax

data SqliteColumnSchemaSyntax

SQLite syntax for column schemas in CREATE TABLE or ALTER COLUMN ... ADD COLUMN statements

Constructors

SqliteColumnSchemaSyntax

Fields

fromSqliteColumnSchema :: SqliteSyntax

sqliteIsSerialColumn :: Bool

data SqliteExpressionSyntax

Implements beam SQL expression syntaxes

Constructors

SqliteExpressionSyntax SqliteSyntax

SqliteExpressionDefault

newtype SqliteValueSyntax

SQLite syntax for values that can be embedded in SqliteSyntax

Constructors

SqliteValueSyntax

Fields

fromSqliteValue :: SqliteSyntax

newtype SqliteTableNameSyntax

Constructors

SqliteTableNameSyntax

Fields

fromSqliteTableName :: SqliteSyntax

newtype SqliteFieldNameSyntax

Constructors

SqliteFieldNameSyntax

Fields

fromSqliteFieldNameSyntax :: SqliteSyntax

newtype SqliteAggregationSetQuantifierSyntax

Constructors

SqliteAggregationSetQuantifierSyntax

Fields

fromSqliteAggregationSetQuantifier :: SqliteSyntax

fromSqliteExpression :: SqliteExpressionSyntax -> SqliteSyntax

SQLite data type syntax

data SqliteDataTypeSyntax

SQLite syntax that implements IsSql92DataTypeSyntax and a good portion of IsSql99DataTypeSyntax, except for array and row types.

Constructors

SqliteDataTypeSyntax

Fields

fromSqliteDataType :: SqliteSyntax

sqliteDataTypeToHs :: HsDataType

sqliteDataTypeSerialized :: BeamSerializedDataType

sqliteDataTypeSerial :: Bool

sqliteTextType :: SqliteDataTypeSyntax

sqliteBlobType :: SqliteDataTypeSyntax

sqliteBigIntType :: SqliteDataTypeSyntax

sqliteSerialType :: SqliteDataTypeSyntax

Building and consuming SqliteSyntax

fromSqliteCommand :: SqliteCommandSyntax -> SqliteSyntax

Convert a SqliteCommandSyntax into a renderable SqliteSyntax

formatSqliteInsert :: SqliteTableNameSyntax -> [Text] -> SqliteInsertValuesSyntax -> SqliteSyntax

Format a SQLite INSERT expression for the given table name, fields, and values.

formatSqliteInsertOnConflict :: SqliteTableNameSyntax -> [Text] -> SqliteInsertValuesSyntax -> Maybe SqliteOnConflictSyntax -> SqliteSyntax

Format a SQLite INSERT expression for the given table name, fields, values, and optionally an ON CONFLICT clause.

emit :: ByteString -> SqliteSyntax

Embed a ByteString directly in the syntax

emitValue :: SQLData -> SqliteSyntax

Emit a properly escaped value into the syntax

This causes a literal ? 3

parens :: SqliteSyntax -> SqliteSyntax

commas :: [SqliteSyntax] -> SqliteSyntax

quotedIdentifier :: Text -> SqliteSyntax

sqliteEscape :: Text -> Text

A best effort attempt to implement the escaping rules of SQLite. This is never used to escape data sent to the database; only for emitting scripts or displaying syntax to the user.

withPlaceholders :: ((SQLData -> Builder) -> Builder) -> Builder

Convert the first argument of SQLiteSyntax to a ByteString Builder, where all the data has been replaced by "?" placeholders.

sqliteRenderSyntaxScript :: SqliteSyntax -> ByteString

Render a SqliteSyntax as a lazy ByteString, for purposes of displaying to a user. Embedded SQLData is directly embedded into the concrete syntax, with a best effort made to escape strings.